api/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.3
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-api-core==2.12.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-pubsub==2.18.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.25.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
